pyproject.toml
[project]
name = "ai-coding-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
	"ruff",
	"pydantic",
	"google-genai",
	"python-dotenv"
]
[tool.ruff]
# Enable standard linting rules (E, F) + imports (I)
select = ["E", "F", "I"]
line-length = 88
ignore=["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
